+++ app/src/app_base.h
@@ -570,6 +570,8 @@ struct st_app_context {
   uint16_t rx_burst_size;
   int force_tx_video_numa;
   int force_rx_video_numa;
+  int force_tx_audio_numa;
+  int force_rx_audio_numa;
 
   bool ptp_systime_sync;
   int ptp_sync_cnt;