+++ app/src/tx_st30p_app.c
@@ -213,6 +213,11 @@ static int app_tx_st30p_init(struct st_app_context* ctx, st_json_st30p_session_t
 
   if (ctx->tx_audio_dedicate_queue) ops.flags |= ST30P_TX_FLAG_DEDICATE_QUEUE;
 
+  if (ctx->force_tx_audio_numa >= 0) {
+    ops.flags |= ST30P_TX_FLAG_FORCE_NUMA;
+    ops.socket_id = ctx->force_tx_audio_numa;
+  }
+
   handle = st30p_tx_create(ctx->st, &ops);
   if (!handle) {
     err("%s(%d), st30p_tx_create fail\n", __func__, idx);